// web/src/app/routes/resume/interfaces/resume.interface.ts
export interface Resume {
  id?: any;
  candidate_id?: number;
  candidate?: string;
  address?: string;
  address_number?: number;
  birthday?: Date;
  email?: string;
  genre?: string;
  hobbies?: any;
  name?: string;
  phone?: number;
  short_description?: any;
  state?: string;
  wage?: number;
  has_resume?: string;
  resume_id?: any;
}